+++ src/data/consultoria-demos.ts
@@ -1,0 +1,9 @@
+/** Demos públicas de consultoría Viento Norte (Figma Make, etc.) */
+
+export const CONSULTORIA_DEMO_X_CMS = {
+  id: "x-cms-n2n",
+  figmaMakeUrl:
+    "https://www.figma.com/make/nHrKYiEtbE0gYnTFB4Ast6/X-%7C-CMS",
+  embedUrl:
+    "https://www.figma.com/embed?embed_host=mi-portafolio&url=https%3A%2F%2Fwww.figma.com%2Fmake%2FnHrKYiEtbE0gYnTFB4Ast6%2FX-%257C-CMS",
+} as const;